Fix async steps in hasStorageAccess and requestStorageAccess.
Fixes #50, #69

This is based on some of Anne's feedback around how we run async steps
in parallel and queue tasks for resolving promises instead of what we
were doing before.

This also moves a bunch of things that probably don't need to happen
in parallel (and don't happen in parallel in at least Gecko and WebKit)
out of the async steps.