Change permission policy feature name to "storage-access" (fixes #104)
This was discussed before in #12 and there was some valid concern around
the "storage-access" name based on the fact that this PP feature is more
focused on "requesting" storage access, and there is no delegation
mechanism like with other permissions that would make it semantically
consistent.

However, I think that in light of #32 and the possibility of integrating
with the permissions API (giving us important functionality such as
observing when storage access is granted) it seems more useful to be
consistent with the (future) permission name and call both "storage-access".