Specify behavior in A > * > A scenarios
This commit tries to make rSA and hSA match the reality of cookie
blocking browsers perform in edge cases such as A > A and A > B > A. I
think all browsers perform a same site check, so moving away from same
origin makes sense. Additionally, Chrome will block cookies based on the
"site for cookies" whereas Firefox and Safari don't. We plan to discuss
this behavior separately and can hopefully align eventually.

As discussed with the other editors, we agree that it still makes sense
from a privacy perspective to automatically grant storage access in a
same site scenario, since the "site for cookies" check is for security
only. This could previously lead to inconsistencies in browsers that
block cookies based on "site for cookies", as hSA would return true and
rSA would resolve automatically without storage access ever being
granted in an ABA case. This commit fixes that by explicitly setting the
"has storage access" bit after a successful same-site check in rSA.

I'm referencing the in-progress "authority" concept on advice of
@annevk.